societywatt-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
python-multipart==0.0.12
pdfplumber==0.11.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
google-generativeai==0.8.3
pillow>=10.4.0
scikit-learn==1.5.2
joblib==1.4.2
numpy>=1.26.4
scipy>=1.13.0
pandas>=2.2.2
psutil>=5.9.8
requests>=2.32.2
bcrypt==4.2.1
PyMuPDF
